feat(sdl): tray, version, video, vulkan branded types

+++ src/sdl/translations/version/functions/index.ts
@@ -1,7 +1,8 @@
 import type { SDL } from '@/sdl';
+import type { Int32 } from '@/types/primitive';
 
 export function getVersion(this: SDL) {
-  return this.symbols.SDL_GetVersion();
+  return this.symbols.SDL_GetVersion() as Int32;
 }
 
 export function getRevision(this: SDL) {
